fixed data type issue

+++ training_scripts/mazi_fixed_grid/test/mfg_t010_f001.py
@@ -1248,7 +1248,7 @@ if __name__=="__main__":
     if model_file is not None:
         with tf.device(tf_device_string):
             model_FANS,training_steps = load_custom()
-            model_FANS.save(savedir+job_name+'_ep'+str(training_steps)+'_model.h5')
+            model_FANS.save(savedir+job_name+'_ep'+str(np.uint(training_steps))+'_model.h5')
             exit()
     else:
         # create a new model
